+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/SeatingPlansApi.java
@@ -111,7 +111,7 @@ public interface SeatingPlansApi {
         produces = { "application/json" }
     )
     default ResponseEntity<SeatingPlanDto> seatingPlansIdGet(
-        @Parameter(name = "id", description = "ID of the seating plan layout that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the seating plan layout that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         getRequest().ifPresent(request -> {
             for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
